@@ -18,6 +18,7 @@
     "extract-text-webpack-plugin": "^0.3.1",
     "fetchival": "^0.3.1",
     "fileapi": "^2.0.10",
+    "fixed-data-table": "^0.4.5",
     "google-maps": "^2.1.1",
     "keymirror": "^0.1.0",
     "moment": "^2.8.3",
@@ -28,19 +29,22 @@
     "react-mixin": "^2.0.2",
     "react-quill": "^0.2.2",
     "react-router": "^0.13.4",
-    "react-select": "git://github.com/taifen/react-select.git#add-onscroll",
+    "react-select": "git://github.com/taifen/react-select.git#custom/AM",
     "react-simpletabs": "^0.6.1",
     "react-sticky": "^2.4.0",
     "react-time-picker": "^1.0.8",
     "react-toastr": "^1.5.2",
     "underscore": "^1.7.0",
-    "whatwg-fetch": "^0.9.0",
+    "uuid": "^2.0.1",
+    "whatwg-fetch": "^0.10.1",
     "when": "^3.7.3"
   },
   "devDependencies": {
-    "babel-core": "^5.8.22",
-    "babel-loader": "^5.3.2",
-    "browser-sync": "^2.9.11",
+    "babel-core": "^5.8.34",
+    "babel-loader": "^5.4.0",
+    "browser-sync": "^2.10.0",
+    "file-loader": "^0.8.5",
+    "fs": "0.0.2",
     "gulp": "^3.8.11",
     "gulp-autoprefixer": "^2.3.1",
     "gulp-clean": "^0.3.1",
@@ -54,13 +58,16 @@
     "gulp-stylus": "^2.1.0",
     "gulp-uglify": "^1.4.2",
     "gulp-util": "^3.0.1",
+    "json-loader": "^0.5.3",
     "gulp-webpack": "^1.4.0",
     "jsx-loader": "^0.13.2",
+    "merge-stream": "^1.0.0",
     "script-loader": "^0.6.0",
     "through2": "^2.0.0",
     "url-loader": "~0.5.4",
-    "webpack": "^1.11.0",
-    "webpack-dev-middleware": "^1.0.11"
+    "webpack": "^1.12.9",
+    "webpack-dev-middleware": "^1.4.0",
+    "webpack-stream": "^2.2.0"
   },
   "scripts": {
     "postinstall": "bower install"
