@@ -1,15 +1,24 @@
-var Flux = require('delorean').Flux;
-var SearchStore = require('../stores/SearchStore');
-var ReportStore = require('../stores/ReportStore');
+import {Flux} from 'delorean';
+import SearchStore from '../stores/SearchStore';
+import ReportStore from '../stores/ReportStore';
+import SearchByTypeStore from '../stores/SearchByTypeStore';
 
-var SearchDispatcher = Flux.createDispatcher({
+export default Flux.createDispatcher({
+
+  setSearchFilters(filter) {
+    return this.dispatch('search:newFilter', filter);
+  },
 
   applySearchFilters(filter) {
     return this.dispatch('search:filter', filter);
   },
 
-  searchResults(filters) {
-    return this.dispatch('search:results', filters);
+  searchResults() {
+    return this.dispatch('search:results');
+  },
+
+  searchResultsByType() {
+    return this.dispatch('search:resultsByType');
   },
 
   getCounters(searchId, query, context) {
@@ -28,12 +37,59 @@ var SearchDispatcher = Flux.createDispatcher({
     this.dispatch('reports:reset');
   },
 
+  initTypeSearch(searchId){
+    return this.dispatch('searchByType:initTypeSearch', searchId);
+  },
+
+  saveTypeSearchModel(model){
+    return this.dispatch('search:setTypeSearchModel', model);
+  },
+
+  setContext(context) {
+    return this.dispatch('searchByType:setContext', context);
+  },
+
+  chooseAssetType(assetType) {
+    return this.dispatch('searchByType:chooseAssetType', assetType);
+  },
+
+  addRow(model) {
+    return this.dispatch('searchByType:addRow', model);
+  },
+
+  addOpenParenthesis(attributes) {
+    return this.dispatch('searchByType:addOpenParenthesis', attributes);
+  },
+
+  addClosingParenthesis(attributes) {
+    return this.dispatch('searchByType:addClosingParenthesis', attributes);
+  },
+
+  deleteRow(model) {
+    return this.dispatch('searchByType:deleteRow', model);
+  },
+
+  moveRowUp(model) {
+    return this.dispatch('searchByType:moveRowUp', model);
+  },
+
+  moveRowDown(model) {
+    return this.dispatch('searchByType:moveRowDown', model);
+  },
+
+  changeRow(model) {
+    return this.dispatch('searchByType:changeRow', model);
+  },
+
+  ensureAttributesLoaded(typeId) {
+    return this.dispatch('searchByType:ensureAttributesLoaded', typeId);
+  },
+
   getStores() {
     return {
+      searchByType: new SearchByTypeStore(),
       results: new SearchStore(),
-      report: new ReportStore(),
+      report: new ReportStore()
     }
   }
 });
-
-module.exports = SearchDispatcher;
