@@ -1,4 +1,5 @@
 module.exports = {
   APIURL: JSON.stringify('http://api.facilitymanager.fmis.software'),
- AUTHURL: JSON.stringify('http://auth.facilitymanager.fmis.software/core'),
+  AUTHURL: JSON.stringify('http://auth.facilitymanager.fmis.software/core'),
+  ADMINURL: JSON.stringify('http://admin.facilitymanager.fmis.software'),
 };
