@@ -2,11 +2,22 @@ var Actions = require('./Actions');
 
 module.exports = class SearchActions extends Actions {
 
+  setSearchFilter(filter) {
+    return this._dispatcher
+      .setSearchFilters(filter);
+  }
+
   changeSearchFilter(filter) {
-    return this._dispatcher.applySearchFilters(filter).then(() => {
-      var results = this._dispatcher.getStore('results').getState();
-      return this._dispatcher.searchResults(results.filter);
-    });
+    return this._dispatcher
+      .applySearchFilters(filter);
+  }
+
+  searchResults() {
+    return this._dispatcher.searchResults();
+  }
+
+  searchResultsByType() {
+    return this._dispatcher.searchResultsByType();
   }
 
   fetchSearchCounters(results) {
