@@ -30,11 +30,14 @@ var GoogleMapsAttributeWrapper = React.createClass({
             return (
                 <div className="asset-data__param" data-param-id={this.props.params.id}>
                     <span className="asset-data__param-title">{this.props.params.name}:</span>
-                    <GoogleMapsComponent
-                        googleMapsApi={this.google.maps}
-                        points={points}
-                        viewOnly={this.props.viewOnly}
-                        onChange={this.handleMapChange} />
+                    <span className="asset-data__param-content">
+                        <GoogleMapsComponent
+                            googleMapsApi={this.google.maps}
+                            points={points}
+                            viewOnly={this.props.viewOnly}
+                            onChange={this.handleMapChange}
+                        />
+                    </span>
                 </div>
             );
         } else {
