@@ -31,7 +31,7 @@ var TypedMoney = React.createClass({
             <ControlWrapper
                 id={this.props.params.id}
                 name={this.props.params.name}
-                className="input-txt input-txt_size_small"
+                className="input-txt input-txt_text"
                 validationState={this.state.validation}
                 isRequired={isRequired}
             >
