@@ -11,10 +11,12 @@ var Attribute = React.createClass({
         return (
             <div className="asset-data__param" data-param-id={params.id}>
                 <span className="asset-data__param-title">{params.name}:</span>
-                <label className="checkbox">
-                    <input type="checkbox" className="checkbox__input" name="checkbox1" checked={value} disabled/>
-                    <span className="checkbox__icon"></span>
-                </label>
+                <span className="asset-data__param-content">
+                    <label className="checkbox">
+                        <input type="checkbox" className="checkbox__input" name="checkbox1" checked={value} disabled/>
+                        <span className="checkbox__icon"></span>
+                    </label>
+                </span>
             </div>
         );
     }
