@@ -10,7 +10,9 @@ var Attribute = React.createClass({
         return (
             <div className="asset-data__param" data-param-id={params.id}>
                 <span className="asset-data__param-title">{params.name}:</span>
-                <a href={"mailto:" + params.value}>{params.value}</a>
+                <span className="asset-data__param-content">
+                    <a href={"mailto:" + params.value}>{params.value}</a>
+                </span>
             </div>
         );
     }
