@@ -10,7 +10,9 @@ var Attribute = React.createClass({
         return (
             <div className="asset-data__param" data-param-id={params.id}>
                 <span className="asset-data__param-title">{params.name}:</span>
-                <span className="no-data">Hidden</span>
+                <span className="asset-data__param-content">
+                    <span className="no-data">Hidden</span>
+                </span>
             </div>
         )
     }
