@@ -6,6 +6,12 @@ var React = require('react');
 var Flux = require('delorean').Flux;
 var Router = require('react-router');
 var Link = Router.Link;
+var L20nMessage = require('../intl/l20n-message');
+
+var searchResultRoutes = {
+    0: 'result', // simple search
+    1: 'resultByType' // searchBytype
+}
 
 var SearchResultsHeader = React.createClass({
 	mixins:[Router.State, Flux.mixins.storeListener],
@@ -23,19 +29,22 @@ var SearchResultsHeader = React.createClass({
     },
     render: function() {
         var tracking = this.state.stores.search.tracking;
-        var query = this.context.router.getCurrentQuery();
+        if (!tracking){
+            return (<div />);
+        }
 
-        return tracking ? (
+        var query = this.context.router.getCurrentQuery();
+        var resultRoute = searchResultRoutes[tracking.searchType];
+        return (
             <div>
-    			<h1 className="page-title">Search results: <span className="page-title__param">{tracking.verboseString}</span></h1>
+    			<h1 className="page-title">{L20nMessage('searchResultsTitle', 'Search results')}: <span className="page-title__param">{tracking.verboseString}</span></h1>
     			<nav className="back-nav">
-                    <Link to="result" params={{searchId : query.searchId}} className="link link_second">
-                        <span className="icon icon_arrow-c_right"></span>Back to search
+                    <Link to={resultRoute} query={{searchId : query.searchId}} className="link link_second">
+                        <span className="icon icon_arrow-c_right"></span>{L20nMessage('backToSearch', 'Back to search')}
                     </Link>
     			</nav>
             </div>
-        )
-        : (<div />);
+        );
     }
 });
 module.exports = SearchResultsHeader;
