@@ -1,7 +1,13 @@
 var React = require('react');
+var Flux = require('delorean').Flux;
+var FixedDataTable = require('fixed-data-table');
 var Loader = require('../common/loader.jsx');
 var LoaderMixin = require('../../mixins/LoaderMixin');
-var Flux = require('delorean').Flux;
+var L20nMessage = require('../intl/l20n-message')
+
+var DataGrid = require('../common/grid');
+
+var Column = FixedDataTable.Column;
 
 var ReportsList = React.createClass({
 
@@ -13,24 +19,66 @@ var ReportsList = React.createClass({
         this.waitFor(this.props.actions.loadReports());
     },
 
+    renderNameCell(cellData) {
+        return <span className="link">{cellData}</span>
+    },
+
+    renderLinkCell(cellData) {
+        return <span className="link"><span className="icon icon_chevron-right"></span></span>
+    },
+
+    renderFinancialCell(cellData) {
+        return cellData ? 'Financial' : ''
+    },
+
+    onRowClick: function(event, rowIndex, data) {
+        var id = data.id;
+        var url = APIURL + '/customreports/index/' + id;
+        var win = window.open(url, '_blank');
+        win.focus();
+    },
+
     render: function() {
         return (
                 <Loader loading={this.state.loading}>
                     <div>
-                        <h1 className="page-title">Reports list</h1>
-                        <table>
-                            <tr>
-                                <th width="80%"></th>
-                                <th width="20%"></th>
-                            </tr>
-                            {this.state.stores.report.reports.map(r => {
-                                var url = APIURL + '/customreports/index/' + r.id;
-                                return <tr>
-                                    <td><a target="_blank" href={url}>{r.name}</a></td>
-                                    <td></td>
-                                </tr>;
-                            })}
-                        </table>
+                        <h1 className="page-title">{L20nMessage('reportsTitle', 'Reports list')}</h1>
+                        <DataGrid
+                            source={this.state.stores.report.reports}
+                            onRowClick={this.onRowClick}
+                            rowHeight={50}
+                            maxHeight={600}
+                            headerHeight={50}
+                            filtering={true}
+                            filterFields={[
+                                {dataKey: 'name', label: 'Name', width: 0.4},
+                                {dataKey: 'assetTypeName', label: 'Asset', width: 0.3},
+                            ]}
+                          >
+                            <Column
+                                label={L20nMessage('reportsColumnName', 'Name')}
+                                width={0.4}
+                                dataKey="name"
+                                cellRenderer={this.renderNameCell}
+                            />
+                            <Column
+                                label={L20nMessage('reportsColumnAsset', 'Asset')}
+                                width={0.3}
+                                dataKey="assetTypeName"
+                            />
+                            <Column
+                                label={L20nMessage('reportsColumnFinancial', 'Financial')}
+                                width={0.25}
+                                dataKey="isFinancial"
+                                cellRenderer={this.renderFinancialCell}
+                            />
+                            <Column
+                                label=" "
+                                width={0.05}
+                                dataKey="id"
+                                cellRenderer={this.renderLinkCell}
+                            />
+                          </DataGrid>
                     </div>
                 </Loader>
             );
