@@ -1,27 +1,31 @@
 import React from 'react';
 import ReactSelectize from '../../common/react-selectize';
-import ValueSelectorBool from './valueSelectorBool';
-import ValueSelectorDate from './valueSelectorDate';
-import ValueSelectorDynList from './valueSelectorDynList';
-import ValueSelectorPlace from './valueSelectorPlace';
-import ValueSelectorRelatedAsset from './valueSelectorRelatedAsset';
-import ValueSelectorText from './valueSelectorText';
+import RowActions from './rowActions';
+import ValueSelectorBool from './valueSelectors/valueSelectorBool';
+import ValueSelectorDate from './valueSelectors/valueSelectorDate';
+import ValueSelectorDynList from './valueSelectors/valueSelectorDynList';
+import ValueSelectorPlace from './valueSelectors/valueSelectorPlace';
+import ValueSelectorText from './valueSelectors/valueSelectorText';
+import ComplexValueSelectorRelatedAsset from './complexValue/ComplexValueSelectorRelatedAsset';
 
 import AssetDispatcher from '../../../dispatchers/AssetDispatcher';
 import AssetActions from '../../../actions/AssetActions';
 
-export default class AttributeRow extends React.Component {
+import SearchDispatcher from '../../../dispatchers/SearchDispatcher';
+import SearchByTypeActions from '../../../actions/SearchByTypeActions';
 
-    logicalOperators = [
-        { name: 'And', id: 1},
-        { name: 'Or', id: 2}
-    ]
+import L20nMessage from '../../intl/l20n-message';
+
+export default class AttributeRow extends React.Component {
 
     constructor(props){
         super(props);
 
         this.assetDispatcher = AssetDispatcher;
         this.assetActions = new AssetActions(this.assetDispatcher);
+
+        this.searchByTypeDispatcher = SearchDispatcher;
+        this.searchByTypeActions = new SearchByTypeActions(this.searchByTypeDispatcher);
     }
 
     onMoveUp = () => {
@@ -41,13 +45,17 @@ export default class AttributeRow extends React.Component {
             return;
         }
 
+        var referenceAttrib = values[0];
+
         var newAttr = this.createNewAttr({
-            referenceAttrib: values[0],
+            referenceAttrib: referenceAttrib,
             operators: [],
             operator: null,
-            value: null
+            value: this.getDefaultValue(referenceAttrib),
+            useComplexValue: false,
+            complexValue: []
         });
-        
+
         this.props.onChange(newAttr);
     }
 
@@ -87,10 +95,50 @@ export default class AttributeRow extends React.Component {
         return Object.assign({}, this.props.selected, diff);
     }
 
+    getDefaultValue(referenceAttrib){
+        var dataType = referenceAttrib.dataType.toLowerCase();
+        var component;
+        switch(dataType){
+            case 'bool':
+                component = ValueSelectorBool;
+                break;
+
+            case 'currentdate':
+            case 'datetime':
+                component = ValueSelectorDate;
+                break;
+
+            case 'dynlist':
+                component = ValueSelectorDynList;
+                break;
+
+            case 'asset':
+            case 'assets':
+            case 'childassets':
+                component = ComplexValueSelectorRelatedAsset;
+                break;
+
+            case 'place':
+                component = ValueSelectorPlace;
+                break;
+
+            default:
+                component = ValueSelectorText;
+        }
+
+        return component.defaultValue || null;
+    }
+
     render(){
-        var valueSelector;
+        var valueSelector = null, complexValueSelector = null;
         var params;
         var dataType = this.props.selected.referenceAttrib.dataType.toLowerCase();
+
+        var logicalOperators = [
+            { name: L20nMessage('searchAnd', 'And'), id: 1},
+            { name: L20nMessage('searchOr', 'Or'), id: 2}
+        ];
+
         switch(dataType){
             case 'bool':
                 valueSelector = <ValueSelectorBool
@@ -116,17 +164,15 @@ export default class AttributeRow extends React.Component {
 
             case 'asset':
             case 'assets':
-                params = {
-                    id: this.props.selected.referenceAttrib.id,
-                    relatedAssetTypeId: this.props.selected.referenceAttrib.relationId,
-                    datatype: this.props.selected.referenceAttrib.dataType,
-                    value: this.props.selected.value || []
-                };
-                valueSelector = <ValueSelectorRelatedAsset
-                    onValueChange={this.onValueChange}
-                    params={params}
-                    dispatcher={this.assetDispatcher}
-                    actions={this.assetActions} />
+            case 'childassets':
+                complexValueSelector = <ComplexValueSelectorRelatedAsset
+                    onChange={this.props.onChange}
+                    selected={this.props.selected}
+                    assetDispatcher={this.assetDispatcher}
+                    assetActions={this.assetActions}
+                    dispatcher={this.searchByTypeDispatcher}
+                    actions={this.searchByTypeActions}
+                    level={this.props.level} />
                 break;
 
             case 'place':
@@ -141,58 +187,56 @@ export default class AttributeRow extends React.Component {
             default:
                 valueSelector = <ValueSelectorText value={this.props.selected.value}
                     onValueChange={this.onValueChange} />;
-        }      
+        }
 
         return (
             <div className="table-search__row">
-                <div className="table-search__row-item table-search__row-item_type_actions">
-                    <span className="table-search__row-action table-search__row-action_delete" title="Delete"
-                        onClick={this.onDelete}></span>
-                    <span className="table-search__row-action table-search__row-action_up" title="Mover up"
-                        onClick={this.onMoveUp}></span>
-                    <span className="table-search__row-action table-search__row-action_down" title="Move down"
-                        onClick={this.onMoveDown}></span>
-                </div>
-                <div className="table-search__row-item table-search__row-item_type_attr">
-                    <ReactSelectize
-                        items={this.props.attributes}
-                        value={this.props.selected.referenceAttrib.id}
-                        onChange={this.onAttrChange}
-                        selectId="selectAttr"
-                        placeholder="Select attribute"
-                        label=" "
-                        clearable={false}
-                        labelField={'displayName'}
-                    />
-                </div>
-                <div className="table-search__row-item table-search__row-item_type_oper">
-                    <ReactSelectize
-                        items={this.props.selected.operators}
-                        value={this.props.selected.operator || 0}
-                        onChange={this.onOperChange}
-                        selectId="selectOperator"
-                        placeholder="Select operator"
-                        label=" "
-                        clearable={false}
-                    />
-                </div>
-                <div className="table-search__row-item table-search__row-item_type_value">
-                    {valueSelector}
-                </div>
-                <div className="table-search__row-item table-search__row-item_type_additional">
-                    <div className={'connector-container ' + (!this.props.selected.lo ? 'hide' : '')}>
-                        <div className="connector">
-                           <ReactSelectize
-                                items={this.logicalOperators}
-                                value={this.props.selected.lo || 1}
-                                onChange={this.onLoChange}
-                                selectId="logicalOperator"
-                                placeholder="Select operator"
-                                label=" "
-                                clearable={false}
-                            />
+                    <RowActions onDelete={this.onDelete} onMoveUp={this.onMoveUp} onMoveDown={this.onMoveDown} />
+                    <div className="table-search__row-item table-search__row-item_type_attr">
+                        <ReactSelectize
+                            items={this.props.attributes}
+                            value={this.props.selected.referenceAttrib.id}
+                            onChange={this.onAttrChange}
+                            selectId="selectAttr"
+                            placeholder="Select attribute"
+                            label=" "
+                            clearable={false}
+                            labelField={'displayName'}
+                        />
+                    </div>
+                    <div className="table-search__row-item table-search__row-item_type_oper">
+                        <ReactSelectize
+                            items={this.props.selected.operators}
+                            value={this.props.selected.operator || 0}
+                            onChange={this.onOperChange}
+                            selectId="selectOperator"
+                            placeholder="Select operator"
+                            label=" "
+                            clearable={false}
+                        />
+                    </div>
+                    <div className="table-search__row-item table-search__row-item_type_value">
+                        {valueSelector}
+                    </div>
+                    <div className="table-search__row-item table-search__row-item_type_additional">
+                        <div className={'connector-container ' + (!this.props.selected.lo ? 'hide' : '')}>
+                            <div className="connector">
+                               <ReactSelectize
+                                    items={logicalOperators}
+                                    value={this.props.selected.lo || 1}
+                                    onChange={this.onLoChange}
+                                    selectId="logicalOperator"
+                                    placeholder="Select operator"
+                                    label=" "
+                                    clearable={false}
+                                />
+                            </div>
                         </div>
                     </div>
+
+                    {complexValueSelector}
+
+                <div className="table-search__row_separator">
                 </div>
             </div>
         );
