@@ -1,25 +1,27 @@
 import React from 'react';
-import ReactSelectize from '../../common/react-selectize';
+import ReactSelectize from '../../../common/react-selectize';
 
 export default class ValueSelectorBool extends React.Component {
-    items = [
-        { name: 'True', id: 1, value: true},
-        { name: 'False', id: 2, value: false}
+    static items = [
+        { name: 'True', id: 1},
+        { name: 'False', id: 2}
     ]
+
+    static defaultValue = { name: 'False', id: 2 }
 
     onValueChange = (items) => {
         if (!items || !items.length){
-            this.props.onValueChange(false);
+            this.props.onValueChange(ValueSelectorBool.defaultValue);
         }
 
-        this.props.onValueChange(items[0].value);
+        this.props.onValueChange(items[0]);
     }
 
     render() {
         return (
             <ReactSelectize
-                items={this.items}
-                value={this.props.value ? 1 : 2}
+                items={ValueSelectorBool.items}
+                value={this.props.value ? this.props.value.id : ValueSelectorBool.defaultValue.id}
                 onChange={this.onValueChange}
                 selectId="selectBoolValue"
                 placeholder="Select value"
