@@ -17,6 +17,8 @@ var LoaderMixin = require('../../mixins/LoaderMixin');
 var Flux = require('delorean').Flux;
 var cx = require('classnames');
 import {param} from "../../util/util";
+import SearchQueryDisplay from './byType/searchQueryDisplay';
+import L20nMessage from '../intl/l20n-message';
 
 var ResultPage = React.createClass({
     mixins: [Router.Navigation, LoaderMixin, Flux.mixins.storeListener],
@@ -26,19 +28,20 @@ var ResultPage = React.createClass({
     contextTypes: {
         router: React.PropTypes.func
     },
-    sortItems: [
-        { name: "Rank", id: 0 },
-        { name: "Date", id: 1 },
-        { name: "Location", id: 2 },
-        { name: "User", id: 3 },
-    ],
+
     exportItems: [
         'txt',
         'xml',
         'xlsx'
     ],
+
+    searchId: 0,
+
+    isSearchByType: function(){
+        return !!this.props.byType;
+    },
+
     getInitialState: function() {
-        var query = this.context.router.getCurrentQuery();
         return {
             searchId: null,
             counters: {
@@ -53,22 +56,42 @@ var ResultPage = React.createClass({
     componentDidMount: function() {
         this.actions = this.props.actions;
         this.props.dispatcher.stores.results.onChange(this.syncUrl);
-        this.loadData(this.context.router.getCurrentQuery());
+
+        var promise = this.actions.setSearchFilter(this.context.router.getCurrentQuery());
+        this.doSearch(promise);
+    },
+
+    componentWillUnmount: function() {
+        this.props.dispatcher.stores.results.listener.removeListener('change', this.syncUrl);
     },
 
     loadData: function(filters, updateCounters = true) {
-        var res = this.actions.changeSearchFilter(filters);
-        this.waitFor(res);
+        var promise = this.actions.changeSearchFilter(filters);
+        this.doSearch(promise, updateCounters);
+    },
+
+    doSearch: function(changeFilterPromise, updateCounters = true){
+        var resultsPromise = changeFilterPromise.then(() => {
+            if (this.isSearchByType()){
+                return this.actions.searchResultsByType();
+            }
+
+            return this.actions.searchResults();
+        });
+
+        this.waitFor(resultsPromise);
+
         if(updateCounters) {
             this.startWaiting('loadingCounters',
-                res.then(() => {
+                resultsPromise.then(() => {
                     this.actions.fetchSearchCounters();
-                    if (filters.assetType) {
-                        this.actions.fetchCustomReportsByType(filters.assetType);
+                    var assetType = this.state.stores.results.filter.assetType;
+                    if (assetType) {
+                        this.actions.fetchCustomReportsByType(assetType);
                     } else {
                         this.actions.resetCustomReports();
                     }
-                }));
+            }));
         }
     },
 
@@ -80,7 +103,8 @@ var ResultPage = React.createClass({
             }
             return acc;
         }, {});
-        this.context.router.transitionTo('/search/result?' + param(clean));
+        var router = this.context.router;
+        router.transitionTo(router.getCurrentPathname() + '?' + param(clean));
     },
 
     filterCounters: function(param, counter) {
@@ -124,6 +148,13 @@ var ResultPage = React.createClass({
         var counters = results.counters;
         var filters = results.filter;
 
+        var sortItems = [
+            { name: L20nMessage('searchResultsSortByRank', 'Rank'), id: 0 },
+            { name: L20nMessage('searchResultsSortByDate', 'Date'), id: 1 },
+            { name: L20nMessage('searchResultsSortByLocation', 'Location'), id: 2 },
+            { name: L20nMessage('searchResultsSortByUser', 'User'), id: 3 },
+        ];
+
         var searchResultsClasses = cx({
             'search-results': true,
             'search-results_type_tiles': this.state.isTilesView,
@@ -143,7 +174,10 @@ var ResultPage = React.createClass({
             'nav-block': true,
         });
 
-        var assetTypeRefinements = counters.assetTypes.filter(this.filterCounters.bind(this, 'assetType'));
+        var assetTypeRefinements = this.isSearchByType()
+            ? []
+            : counters.assetTypes.filter(this.filterCounters.bind(this, 'assetType'));
+
         var taxonomyRefinements = counters.taxonomies.filter(this.filterCounters.bind(this, 'taxonomy'));
 
         var postsPerPage = 20;
@@ -155,20 +189,40 @@ var ResultPage = React.createClass({
         var urlQuery = this.context.router.getCurrentQuery();
         var isHistory = urlQuery.context == 2;
 
+        var assetTypeName, attributes;
+        if (!this.state.stores.results.searchByTypeModel){
+            assetTypeName = '';
+            attributes = [];
+        } else {
+            assetTypeName = this.state.stores.results.searchByTypeModel.assetType.name;
+            attributes = this.state.stores.results.searchByTypeModel.attributes;
+        }
+
         return (
             <div>
-                <div className="grid">
+                <div className="grid results-grid">
                     <div className="grid__item two-twelfths">
-                        <h1 className="page-title page-title_small">Search results</h1>
+                        <h1 className="page-title page-title_small">{L20nMessage('searchResultsTitle', 'Search results')}</h1>
                     </div>
                     <div className="grid__item ten-twelfths">
-                        {this.context.router.getCurrentQuery().query
+                        {!this.isSearchByType()
                           ?  <SearchSimpleForm
                                 dispatcher={this.props.dispatcher}
                                 changeFilter={this.loadData}
                                 value={urlQuery.query}
                                 context={urlQuery.context} />
-                          : {}
+                          : <div>
+                                <div>
+                                    <SearchQueryDisplay
+                                        typeName={assetTypeName}
+                                        attributes={attributes} />
+                                </div>
+                                <div>
+                                    <Link to="type-search" query={{ searchId: urlQuery.searchId }}>
+                                        edit search
+                                    </Link>
+                                </div>
+                            </div>
                         }
                     </div>
                 </div>
@@ -190,9 +244,9 @@ var ResultPage = React.createClass({
                                         </button>
                                     </div>
                                     <span className="input-group__item">
-                                        <span className="input-group__item-title">Sort by</span>
+                                        <span className="input-group__item-title">{L20nMessage('searchResultsSort', 'Sort by')}</span>
                                         <ReactSelectize
-                                            items={this.sortItems}
+                                            items={sortItems}
                                             value={filters.sortBy}
                                             onChange={this.handleSortChange}
                                             selectId="select-sortby"
@@ -219,7 +273,7 @@ var ResultPage = React.createClass({
                                 <div>
                                     {assetTypeRefinements.length !== 0
                                         ? <RefinementBlock
-                                            title="Refine by assets"
+                                            title={L20nMessage('searchResultsRefineAssets', 'Refine by assets')}
                                             list={assetTypeRefinements}
                                             type="assetType"
                                             changeFilter={self.loadData}
@@ -230,7 +284,7 @@ var ResultPage = React.createClass({
                                     }
                                     {taxonomyRefinements.length !== 0
                                         ? <RefinementBlock
-                                            title="Refine by taxonomies"
+                                            title={L20nMessage('searchResultsRefineTax', 'Refine by taxonomies')}
                                             list={taxonomyRefinements}
                                             type="taxonomy"
                                             changeFilter={self.loadData}
@@ -243,7 +297,7 @@ var ResultPage = React.createClass({
                                         <ReportsBlock searchId={results.searchId} reports={this.state.stores.report.reports} />
                                     </nav>
                                     <nav className={navBlockClasses}>
-                                        <span className="nav-block__title">Export</span>
+                                        <span className="nav-block__title">{L20nMessage('searchResultsExport', 'Export')}</span>
                                         <ul className="nav-block__list">
                                             {this.exportItems.map(function(format){
                                                 return <li className="nav-block__item">
@@ -261,10 +315,10 @@ var ResultPage = React.createClass({
                             <Loader loading={this.state.loading} >
                                 <div className={searchResultsClasses}>
                                     <header className="search-results__header">
-                                        <span className="search-results__header-item search-results__header-item_name">Assets path</span>
-                                        <span className="search-results__header-item search-results__header-item_category">Category</span>
-                                        <span className="search-results__header-item search-results__header-item_attr">Attributes set</span>
-                                        <span className="search-results__header-item search-results__header-item_link">Go to result</span>
+                                        <span className="search-results__header-item search-results__header-item_name">{L20nMessage('searchResultsColumnName', 'Name')}</span>
+                                        <span className="search-results__header-item search-results__header-item_category">{L20nMessage('searchResultsColumnCategory', 'Category')}</span>
+                                        <span className="search-results__header-item search-results__header-item_attr">{L20nMessage('searchResultsColumnAttr', 'Attributes set')}</span>
+                                        <span className="search-results__header-item search-results__header-item_link">{L20nMessage('searchResultsColumnLink', 'Go to result')}</span>
                                     </header>
                                     <ul className="search-results__list">
                                         {results.models.length !== 0
@@ -278,7 +332,7 @@ var ResultPage = React.createClass({
                                             : {}
                                         }
                                         {results.models.length === 0 && this.state.loading === false
-                                            ? <li className="search-results__item search-results__item_empty"><span className="search-results__item-msg"><strong>Nothing was found</strong> for the specified search parameters</span></li>
+                                            ? <li className="search-results__item search-results__item_empty"><span className="search-results__item-msg">{L20nMessage('searchResultsNotFound', 'Nothing was found for the specified search parameters')}</span></li>
                                             : {}
                                         }
                                     </ul>
