@@ -21,17 +21,24 @@ var ResultItem = React.createClass({
             <li className="search-results__item">
                 <a className="search-results__item-link" href={assetLink}>
                     <span className="search-results__item-param search-results__item-param_name">
-                        <span className="link">{this.props.model.displayValues}</span>
+                        <span className="link">{this.props.model.name}</span>
                     </span>
                     <span className="search-results__item-param search-results__item-param_category">
-                        <span className="label">{this.props.model.categoryKeywords}</span>
+                        {this.props.model.categoryKeywords
+                            ? <span className="label">{this.props.model.categoryKeywords}</span>
+                            : null
+                        }
                     </span>
                     <span className="search-results__item-param search-results__item-param_attr">
                         <span className="search-results__item-attr">
-                            {this.props.model.allAttrib2IndexValues}
+                            {this.props.model.displayValues.map((item) => {
+                                return <p><strong>{item.key}:</strong> {item.value}</p>
+                            })}
                         </span>
-                    <span className="search-results__item-attr search-results__item-attr_extended">
-                            {this.props.model.allAttribValues}
+                        <span className="search-results__item-attr search-results__item-attr_extended">
+                            {this.props.model.displayExtValues.map((item) => {
+                                return <p><strong>{item.key}:</strong> {item.value}</p>
+                            })}
                         </span>
                     </span>
                     <span className="search-results__item-param search-results__item-param_link">
