@@ -1,10 +1,11 @@
-var LoginStore = require('../stores/LoginStore.js');
+var LoginStore = require('../stores/LoginStore.js').store;
 
 var AuthenticatedRouteMixin = {
     statics: {
         willTransitionTo: function (transition, params, query) {
-            if (!LoginStore.store.isLoggedIn()) {
+            if (!LoginStore.isLoggedIn()) {
               localStorage['nextPath'] = transition.path;
+              transition.redirect('login');
             } else if (localStorage['nextPath']) {
               let nextPath = localStorage['nextPath'];
               localStorage['nextPath'] = '';
