@@ -2,21 +2,28 @@ import fetch from "../util/fetch";
 
 export default class SearchRepository {
     search(params) {
-        var searchUrl;
-        if (!params.query && params.assetType){
-            searchUrl = '/api/search/bytype';
-        } else {
-            searchUrl = '/api/search';
-        }
-        
-        return fetch(searchUrl).get({
+        return fetch('/api/search').get({
             searchId: params.searchId,
             query: params.query,
             page: params.page,
             assetType: params.assetType,
             taxonomy: params.taxonomy,
             sortBy: params.sortBy,
-            context: params.context
+            context: params.context,
+            attributeId: params.attributeId,
+            assetId: params.assetId
+        });
+    }
+
+    searchByType(params, model) {
+        return fetch('/api/search/bytype').post({
+            searchId: model.searchId,
+            assetType: model.assetType,
+            assetTypeContext: model.assetTypeContext,
+            attributes: model.attributes,
+            page: params.page,
+            taxonomy: params.taxonomy,
+            sortBy: params.sortBy
         });
     }
 
